+++ .storybook/webpack.config.js
@@ -1,3 +1,5 @@
+const path = require('path');
+
 module.exports = {
   module: {
     rules: [
@@ -12,6 +14,15 @@ module.exports = {
             options: {
               sourceMap: true
             }
+          }
+        ]
+      },
+      {
+        test: /\.(png|jpg|gif|svg)$/,
+        use: [
+          {
+            loader: 'file-loader',
+            options: {}
           }
         ]
       }