+++ src/index.js
@@ -1,6 +1,5 @@
 module.exports = {
   HTML: require('./html'),
   Text: require('./text'),
-  Editor: require('./editor'),
-  Container: require('./container')
+  Editor: require('./editor')
 };